requirements.txt
requests==2.26.0
Django==3.2.7
djangorestframework==3.13.1
psycopg2==2.9.2
django-q==1.3.9